Dropped length normalization support
It is just theorically wrong. We assume sentence A is better than B
when P(A) is larger than P(B). That means LogP(A) > LogP(B) or
-LogP(A) < -LogP(B) hence \sigma edge(A) < \sigma edge(B).
There is nothing to normalize here. If B has more words, it has to
have larger probability in order to beat A, that's it.