Reworked WordArchive::load to use new wordlist format
The new format is as same as CMU SLM's vocab format. It should
be noted that words in wordlist are standardized ones.

If WordArchive::load() is called with NULL as argument, it'll
then use wordlist from struct lm_t inside class LM. So if you
already load an LM, call warch.load(NULL) to save I/O.